Fetch the results when the location changes
