+++ src/app/query-home/flows/new-query-version.ts
@@ -2,9 +2,6 @@ import {nanoid} from "@reduxjs/toolkit"
 import {Thunk} from "src/js/state/types"
 import QueryVersions, {QueryVersion} from "src/js/state/QueryVersions"
 import {getQuerySource} from "./get-query-source"
-import tabHistory from "../../router/tab-history"
-import {lakeQueryPath} from "../../router/utils/paths"
-import Current from "src/js/state/Current"
 import RemoteQueries from "src/js/state/RemoteQueries"
 import {setRemoteQueries} from "../../features/sidebar/flows/remote-queries"
 
@@ -14,7 +11,6 @@ export const newQueryVersion =
     attrs: Partial<QueryVersion> = {}
   ): Thunk<Promise<QueryVersion>> =>
   async (dispatch, getState) => {
-    const lakeId = Current.getLakeId(getState())
     const version: QueryVersion = {
       value: "",
       ...attrs,
@@ -29,7 +25,6 @@ export const newQueryVersion =
     }
 
     dispatch(QueryVersions.add({queryId, version}))
-    dispatch(tabHistory.push(lakeQueryPath(queryId, lakeId)))
 
     return version
   }
