+++ src/app/query-home/flows/submit-search.ts
@@ -3,6 +3,8 @@ import Editor from "src/js/state/Editor"
 import Notice from "src/js/state/Notice"
 import {newQueryVersion} from "./new-query-version"
 import Results from "src/js/state/Results"
+import tabHistory from "src/app/router/tab-history"
+import {lakeQueryPath} from "src/app/router/utils/paths"
 
 type SaveOpts = {history: boolean; investigation: boolean}
 
@@ -12,6 +14,7 @@ const submitSearch =
     _ts: Date = new Date()
   ) =>
   (dispatch, getState) => {
+    console.log("submit search")
     dispatch(Notice.dismiss())
     dispatch(Results.error(null))
     const value = Editor.getValue(getState())
@@ -26,11 +29,8 @@ const submitSearch =
         dispatch(Results.error(error))
         return
       }
-      // TODO: Mason - refactor history to use query copies
-      // if (save.investigation) {
-      //   dispatch(Investigation.push(lakeId, poolId, record, brim.time(ts).toTs()))
-      // }
-      dispatch(Results.fetchFirstPage(query.toString()))
+      const lakeId = Current.getLakeId(getState())
+      dispatch(tabHistory.push(lakeQueryPath(query.id, lakeId)))
     })
   }
