+++ src/app/query-home/index.tsx
@@ -1,9 +1,8 @@
 import {useExpandState} from "src/app/query-home/results/expand-hook"
-import {DRAFT_QUERY_NAME} from "src/app/query-home/utils/brim-query"
 import {ActionButtonProps} from "src/app/toolbar/action-button"
 import Layout from "src/js/state/Layout"
 import ResultsComponent from "./results"
-import React, {useLayoutEffect} from "react"
+import React from "react"
 import {useSelector} from "react-redux"
 import {useDispatch} from "src/app/core/state"
 import Current from "src/js/state/Current"
@@ -12,59 +11,15 @@ import Toolbar from "./toolbar"
 import styled from "styled-components"
 import useExport from "./toolbar/hooks/use-export"
 import useColumns from "./toolbar/hooks/use-columns"
-import DraftQueries from "src/js/state/DraftQueries"
-import {syncPool} from "../core/pools/sync-pool"
 import ToolbarButton from "./toolbar/button"
 import {newDraftQuery} from "./flows/new-draft-query"
 import tabHistory from "../router/tab-history"
 import {lakeQueryPath} from "../router/utils/paths"
-import {getQuerySource} from "./flows/get-query-source"
 import SearchArea from "./search-area"
 import {FeatureFlag} from "../core/feature-flag"
 import RightPane from "../features/right-pane"
 import usePins from "./toolbar/hooks/use-pins"
-import Editor from "src/js/state/Editor"
 import {usePinContainerDnd} from "./search-area/pins/use-pin-dnd"
-import Results from "src/js/state/Results"
-import submitSearch from "./flows/submit-search"
-
-const syncQueryLocationWithRedux = (dispatch, getState) => {
-  const {queryId} = Current.getQueryLocationData(getState())
-  const lakeId = Current.getLakeId(getState())
-  const query = Current.getQuery(getState())
-  const isDraft = dispatch(getQuerySource(queryId)) === "draft"
-  const pool = Current.getQueryPool(getState())
-  if (pool && !pool.hasSpan()) dispatch(syncPool(pool.id, lakeId))
-  if (!query && queryId && isDraft) {
-    // reset drafts?
-    dispatch(
-      DraftQueries.set({
-        id: queryId,
-        name: DRAFT_QUERY_NAME,
-        value: "",
-        pins: {},
-      })
-    )
-  }
-
-  // Give codemirror a chance to update by scheduling this update
-  setTimeout(() => {
-    dispatch(Editor.setValue(query?.value || ""))
-    dispatch(Editor.setPins(query?.pins || []))
-    if (Results.getStatus(getState()) === "INIT") {
-      dispatch(submitSearch())
-    }
-  })
-}
-
-export function useSearchParamLocationSync() {
-  const dispatch = useDispatch()
-  const location = useSelector(Current.getLocation)
-
-  useLayoutEffect(() => {
-    dispatch(syncQueryLocationWithRedux)
-  }, [location.key])
-}
 
 const QueryPageHeader = styled.div`
   background: white;
@@ -136,7 +91,6 @@ const ContentWrap = styled.div`
 `
 
 const QueryHome = () => {
-  useSearchParamLocationSync()
   const query = useSelector(Current.getQuery)
   const lakeId = useSelector(Current.getLakeId)
   const drop = usePinContainerDnd()
